% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_dist.R
\name{acd}
\alias{acd}
\title{Absolute coverage difference (ACD)}
\usage{
acd(actual, lower, upper, level)
}
\arguments{
\item{actual}{actual values (Equivalent of upper/lower)}

\item{lower}{lower bound of prediction interval}

\item{upper}{upper bound of prediction interval}

\item{level}{level used for prediction interval construction}
}
\value{
numeric vector of length 1
}
\description{
Absolute coverage difference (ACD)
}
\references{
\itemize{
\item Makridakis, S., Spiliotis, E., & Assimakopoulos, V. (2020). The M4
Competition: 100,000 time series and 61 forecasting methods. International
Journal of Forecasting, 36(1), 54–74. \url{https://doi.org/10.1016/j.ijforecast.2019.04.014}
}
}
\seealso{
Other PI accuracy measures: 
\code{\link{smis}()}
}
\concept{PI accuracy measures}
