% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_diffs}
\alias{add_diffs}
\title{Add differences to time series data.table object for specified columns}
\usage{
add_diffs(DT, cols, ndiff)
}
\arguments{
\item{DT}{data.table object with indexed time variable}

\item{cols}{columns to use for differencing}

\item{ndiff}{number of lags}
}
\value{
nothing returned but data.table edited by reference
}
\description{
Add differences to time series data.table object for specified columns
}
