% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_growth_rates}
\alias{add_growth_rates}
\title{Add growth rates to time series data.table object for specified columns}
\usage{
add_growth_rates(DT, cols, ndiff)
}
\arguments{
\item{DT}{data.table object with indexed time variable}

\item{cols}{columns to use for growth rates}

\item{ndiff}{number of lags}
}
\value{
nothing returned but data.table edited by reference
}
\description{
Add growth rates to time series data.table object for specified columns
}
