% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_shift}
\alias{add_shift}
\title{Add differences to time series data.table object for specified columns}
\usage{
add_shift(DT, cols, nlags, type = "lag")
}
\arguments{
\item{DT}{data.table object with indexed time variable}

\item{cols}{columns to use for shifting (character)}

\item{nlags}{number of lags}

\item{type}{one of "lag", "lead" and "shift"}
}
\value{
nothing returned, DT edited by reference
}
\description{
Add differences to time series data.table object for specified columns
}
