% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_fcf}
\alias{apple_fcf}
\title{Quarterly reported free cash flow from Apple as ts object from 1995 to 2020.}
\format{
A quarterly ts object with from 1995 to 2020
}
\usage{
data(apple_fcf)
}
\description{
Quarterly reported free cash flow from Apple as ts object from 1995 to 2020.
}
\keyword{datasets}
