% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_acf.R
\name{check_acf}
\alias{check_acf}
\title{Check for autocorrelation}
\usage{
check_acf(data, level = 0.05, lag_grid = c(8, 12, 16))
}
\arguments{
\item{data}{a numeric vector or univariate time series object}

\item{level}{significance level}

\item{lag_grid}{vector of lags to test on}
}
\value{
list with data.frames of p-values for both \code{Ljung-Box} and
\code{Box-Pierce} tests results and logical vector \code{pass} which is TRUE if all
tests passes and FALSE otherwise
}
\description{
Ljung-Box tests for 8, 12 and 16 lags
}
\examples{
apple <- fcf::ts_apple

# With all default
check_acf(apple)

# Significance level to 0.01 and 1 to 5 lags
check_acf(apple, level = 0.01, lag_grid = 1:5)

}
