% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dow30_clean}
\alias{dow30_clean}
\title{US index Dow Jones 30 quarterly EBIT from 1990 to 2020, cleaned by missing
data and outliers}
\format{
A data.table with 2880 rows and 3 variables (xz-compressed):
\describe{
\item{ticker}{bbreviation of company title}
\item{index}{Publishing date of (quarterly) report in POSIX format.
Month and year of interest}
\item{value}{Earnings before interest and taxes}
}
}
\usage{
data(dow30_clean)
}
\description{
US index Dow Jones 30 quarterly EBIT from 1990 to 2020, cleaned by missing
data and outliers
}
\keyword{datasets}
