% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_predict.R
\name{dropout_predict}
\alias{dropout_predict}
\title{predict for keras models}
\usage{
dropout_predict(model, x)
}
\arguments{
\item{model}{keras model}

\item{x}{array (with shape similar to input)}
}
\value{
keras array
}
\description{
\code{py_suppress_warnings} wrapped around predict
}
