% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT_apple}
\alias{DT_apple}
\title{Quarterly reported free cash flow from Apple as data.table object from 1995
to 2020.}
\format{
A quarterly data.table object from 1995 to 2020
}
\usage{
data(DT_apple)
}
\description{
Quarterly reported free cash flow from Apple as data.table object from 1995
to 2020.
}
\keyword{datasets}
