% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT_unh}
\alias{DT_unh}
\title{Quarterly reported EBIT from UNH as data.table object from 1995
to 2020.}
\format{
A quarterly data.table object from 1995 to 2020
}
\usage{
data(DT_unh)
}
\description{
Quarterly reported EBIT from UNH as data.table object from 1995
to 2020.
}
\keyword{datasets}
