% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_keras_sequential.R
\name{evaluate_keras_sequential}
\alias{evaluate_keras_sequential}
\title{Evaluation Measures for keras sequential model forecasts}
\usage{
evaluate_keras_sequential(DT, forecast, n_train, n_val, n_test)
}
\arguments{
\item{DT}{actual data (including train, validation and test set)}

\item{forecast}{forecast values}

\item{n_train}{length of training set}

\item{n_val}{length of validation set}

\item{n_test}{length of test set}
}
\value{

}
\description{
Evaluation Measures for keras sequential model forecasts
}
