% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fc_arima}
\alias{fc_arima}
\title{Univaritate time series with actual datapoints and ARIMA forecast (data.table
object)}
\format{
A data.table with 58 rows and 5 variables:
\describe{
\item{index}{Publishing date of (quarterly) report}
\item{value}{Earnings before interest and taxes}
\item{key}{"actual" and "predict"}
\item{lo95}{95\% confidence interval lower barrier}
\item{hi95}{95\% confidence interval upper barrier}
}
}
\usage{
data(fc_arima)
}
\description{
Univaritate time series with actual datapoints and ARIMA forecast (data.table
object)
}
\keyword{datasets}
