% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fc_baseline}
\alias{fc_baseline}
\title{Baseline Forecast (data.table object) with SNAIVE and Holt's Trend}
\format{
A quarterly data.table object from 2000 to 2004
}
\usage{
data(fc_baseline)
}
\description{
Baseline Forecast (data.table object) with SNAIVE and Holt's Trend
}
\keyword{datasets}
