% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_baseline.R
\name{forecast_baseline}
\alias{forecast_baseline}
\title{baseline models to get benchmark results for advanced models}
\usage{
forecast_baseline(data, h = NULL, acc_measure)
}
\arguments{
\item{data}{univariate time series object}

\item{h}{atomic vector specifying forecasting horizon}

\item{acc_measure}{Accuracy indicator. Valid options are "ME", "RMSE", "MAE",
"MPE", "MAPE", "MASE" and "ACF1"}
}
\value{

}
\description{
Get results for the following simple forecasting methods:
\itemize{
\item naive forecast
\item seasonal naive forecast
\item mean forecast
\item Exponential smoothing forecasts (\link[forecast]{ses})
\item holt
}
}
