% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_fcf}
\alias{get_fcf}
\title{Calculate Free Cash Flows from Financial Data}
\usage{
get_fcf(df_in)
}
\arguments{
\item{df_in}{Prepped data frame as returned by \code{\link[=prep_financials]{prep_financials()}}.}
}
\description{
Calculate Free Cash Flows from Financial Data
}
