% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_prices}
\alias{get_prices}
\title{Get historical price data}
\usage{
get_prices(sym)
}
\arguments{
\item{sym}{The stock ticker symbol of interest, passed as a \code{string}.}
}
\description{
Get historical price data
}
