% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_basic_sequential.R
\name{keras_basic_sequential}
\alias{keras_basic_sequential}
\title{Basic sequential NN by Keras framework using \link\href{flags}{tfruns}}
\usage{
keras_basic_sequential(X, Y)
}
\arguments{
\item{X}{list of "train", "val", and "test" with 3D (keras) arrays}

\item{Y}{list of "train", "val", and "test" with 2D (keras) arrays}
}
\value{
history
}
\description{
Basic sequential NN by Keras framework using \link\href{flags}{tfruns}
}
\section{FLAGS}{

The following parameters are currently required:
\itemize{
\item optimizer
\item loss
\item metrics
\item patience
}
}

