% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_dist.R
\name{acd}
\alias{acd}
\alias{smis}
\title{Absolute coverage difference (ACD)}
\usage{
acd(actual, lower, upper, level)

smis(data, lower, upper, h, m, level)
}
\arguments{
\item{actual}{actual values (only test set)}

\item{lower}{lower bound of prediction interval}

\item{upper}{upper bound of prediction interval}

\item{level}{level used for prediction interval construction}

\item{data}{actual values (containing both train and evaluation values)}

\item{h}{forecast horizon}

\item{m}{frequency, e.g. 12 for monthly and 4 for quarterly series}
}
\value{
vector of length 1

vector of length 1
}
\description{
Absolute coverage difference (ACD)

Scaled Mean Interval Score, scaled according to M4 Forecasting competition
}
\references{
\itemize{
\item Makridakis, S., Spiliotis, E., & Assimakopoulos, V. (2020). The M4
Competition: 100,000 time series and 61 forecasting methods. International
Journal of Forecasting, 36(1), 54–74. \url{https://doi.org/10.1016/j.ijforecast.2019.04.014}
}

\itemize{
\item \link[greybox]{sMIS}
\item Gneiting, T., & Raftery, A. E. (2007). Strictly proper scoring rules,
prediction, and estimation. Journal of the American Statistical Association,
102(477), 359–378. \url{https://doi.org/10.1198/016214506000001437}
\item Makridakis, S., Spiliotis, E., & Assimakopoulos, V. (2020). The M4
Competition: 100,000 time series and 61 forecasting methods. International
Journal of Forecasting, 36(1), 54–74. \url{https://doi.org/10.1016/j.ijforecast.2019.04.014}
}
}
