% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_point.R
\name{mase}
\alias{mase}
\alias{mape}
\alias{smape}
\title{Point Forecast Accuracy Measures}
\usage{
mase(data, forecast, m)

mape(actual, forecast)

smape(actual, forecast)
}
\arguments{
\item{data}{actual values (containing both train and evaluation values)}

\item{forecast}{forecasted values}

\item{m}{frequency, e.g. 12 for monthly and 4 for quarterly series}

\item{actual}{actual values (only test set)}
}
\value{
vector of length 1
}
\description{
Point Forecast Accuracy Measures
}
\section{mase}{

Mean Absolute Scaled Error (MASE), scaled according to M4 Forecasting
competition
}

\section{mape}{

Mean Absolute Percentage Error (MAPE)
}

\section{smape}{

symmetric Mean Absolute Percentage Error (sMAPE)
}

\references{
\itemize{
\item 	Svetunkov, I., Sagaert, Y. R. (2020). greybox: Toolbox for Model
Building and Forecasting. \url{https://cran.r-project.org/package=greybox}
\item Makridakis, S., Spiliotis, E., & Assimakopoulos, V. (2020). The M4
Competition: 100,000 time series and 61 forecasting methods. International
Journal of Forecasting, 36(1), 54–74. \url{https://doi.org/10.1016/j.ijforecast.2019.04.014}
}
}
