% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tf_version.R
\name{parse_tf_version}
\alias{parse_tf_version}
\title{Parse installed TensorFlow version}
\usage{
parse_tf_version()
}
\value{
numeric vector length 1
}
\description{
Parse installed TensorFlow version
}
\examples{
parse_tf_version()
}
