% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baselines.R
\name{plot_baselines}
\alias{plot_baselines}
\title{Plot Baseline forecasts}
\usage{
plot_baselines(
  data,
  col_date = "index",
  col_value = "value",
  col_group = "type",
  title = NULL,
  size = 0.6,
  alpha = 0.8,
  colors = NULL,
  legend = "bottom",
  scale = NULL
)
}
\arguments{
\item{data}{data.frame containing "Date", "value" and "group"}

\item{col_date}{Date column in \code{data}, default to "index"}

\item{col_value}{Value column in \code{data}, default to "value"}

\item{col_group}{Group column in \code{data}, default to "type" (e.g.
"Actual", "Naive", ...) - see example}

\item{title}{diagram title}

\item{size}{line size}

\item{alpha}{line alpha value}

\item{colors}{named character vector to fit color to type}

\item{legend}{legend position}

\item{scale}{NULL or two Date values to scale x-axis}
}
\value{
ggplot2 object
}
\description{
Plot results of \link{baselines} (by \code{ggplot2})
}
\examples{
plot_baselines(
  data = tsRNN::fc_baseline,
  colors = c("Actual" = "black", "Snaive" = "blue", "Holt" = "green")
)
}
