% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot time series (incl. forecasts) for single cross validation split}
\usage{
plot_prediction(
  data,
  col_date = "index",
  col_value = "value",
  col_group = "key",
  title = NULL,
  size = 0.6,
  alpha = 0.8,
  legend = "bottom",
  scale = NULL,
  PI = FALSE,
  col_pi_high = "hi95",
  col_pi_low = "lo95"
)
}
\arguments{
\item{data}{data.frame containing "index" (Date), "value" (numeric) and key
(character - "actual" and "predict")}

\item{col_date}{Date column in \code{data}, default to "index"}

\item{col_value}{Value column in \code{data}, default to "value"}

\item{col_group}{Group column in \code{data}, default to "key" ("actual" and
"predict") - see example}

\item{title}{plot title}

\item{size}{line size}

\item{alpha}{line alpha value}

\item{legend}{legend position}

\item{scale}{NULL or two Date values to scale x-axis}

\item{PI}{add prediction interval?}

\item{col_pi_high}{Upper confidence bound in \code{data}}

\item{col_pi_low}{Lower confidence bound in \code{data}}
}
\value{
ggplot2 object
}
\description{
E.g. plot result for selected split from \link{cv_arima} (by \code{ggplot2})
}
\examples{
plot_prediction(
  data = tsRNN::fc_arima,
  title = "ARIMA Forecast versus actual data"
)

# With Prediction Interval
plot_prediction(
  data = tsRNN::fc_arima,
  title = "ARIMA Forecast with prediction interval",
  PI = TRUE
)
}
