% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction_samples}
\alias{plot_prediction_samples}
\title{Plot multiple splits from same \code{rsample} split}
\usage{
plot_prediction_samples(
  splits,
  col_date = "index",
  col_value = "value",
  col_group = "key",
  title = NULL,
  date_type = "datetime",
  ncol = 3L,
  scale = NULL,
  PI = FALSE,
  col_pi_high = "hi95",
  col_pi_low = "lo95"
)
}
\arguments{
\item{splits}{list of prediction data.frames}

\item{title}{diagram title}

\item{date_type}{string vector - one of "datetime", "date" or "character"}

\item{ncol}{Number of columns}

\item{scale}{NULL or two Date values to scale x-axis}

\item{PI}{add prediction interval?}
}
\value{
ggplot2 object
}
\description{
Plot multiple splits from same \code{rsample} split
}
\examples{
data <- fcf::fc_arima
plot_prediction_samples(
   splits = list(data, data),
   ncol = 2L
)

# with Prediction Interval
plot_prediction_samples(
  splits = list(data, data),
  ncol = 2L,
  PI = TRUE
)
}
