% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_arima.R
\name{predict_arima}
\alias{predict_arima}
\title{Forecast the next n steps by ARIMA for each split}
\usage{
predict_arima(
  data,
  cv_setting,
  transform = NULL,
  frequency = 4,
  multiple_h = NULL
)
}
\arguments{
\item{data}{data.table object}

\item{cv_setting}{cross valudation settings. List Requiring \itemize{
\item periods_train
\item periods_val
\item periods_test
\item skip_span
}}

\item{transform}{One of NULL (default) and "normalize"}

\item{frequency}{time series frequency, e.g. 4 for quarters and 12 for months}

\item{multiple_h}{NULL if forecast horizon equals cv_setting$n_test, else
named list of forecast horizons for accuracy measures}
}
\value{
list with accuracy and forecasts (point forecast and PI)
}
\description{
Forecast the next n steps by ARIMA for each split
}
