% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_keras_rnn.R
\name{predict_keras_rnn}
\alias{predict_keras_rnn}
\title{Title}
\usage{
predict_keras_rnn(
  DT,
  model_type = "simple",
  multiple_h = NULL,
  epochs = 200,
  lag_setting = 1:4,
  length_val = 8,
  length_test = 8,
  n_units = 32,
  optimizer_type = "rmsprop",
  save_model = FALSE,
  filepath = NULL,
  forecast_future = FALSE,
  forecast_length = NULL,
  ...
)
}
\arguments{
\item{DT}{univariate time series - data.table object with 2 variables:
"index" (Posixct date-time) and "value". Usually single DT from \code{rsample}}

\item{model_type}{One of "simple", "gru" and "lstm"}

\item{multiple_h}{NULL if forecast horizon equals cv_setting$n_test, else
named list of forecast horizons for accuracy measures}

\item{epochs}{number of epochs to train model (default to 300)}

\item{lag_setting}{numeric vector specifying lags to train on}

\item{length_val}{length of validation set}

\item{length_test}{length of test set}

\item{n_units}{32 (currently fixed)}

\item{optimizer_type}{One of "rmsprop" (default) and "adam"}

\item{save_model}{save model? If TRUE, \code{file_path} required}

\item{filepath}{path to save model. Must end with \code{.hdf5}}

\item{forecast_future}{produce forecasts?}

\item{forecast_length}{of \code{forecast_future}, specify forecast horizon}

\item{...}{additional parameters passed to Keras}
}
\value{
Predictions
}
\description{
Title
}
\examples{
\dontrun{

# Prepare data
apple <- tsRNN::dow30[
  ticker == "AAPL" & date > as.POSIXct("1995-01-01"),
  .SD, .SDcols = c("date", "fcf")
]
data.table::setnames(apple, c("date", "fcf"), c("index", "value"))

# Predict
predict_keras_rnn(DT = apple)

}
}
