% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_keras_sequential.R
\name{predict_keras_sequential}
\alias{predict_keras_sequential}
\title{Title}
\usage{
predict_keras_sequential(
  DT,
  epochs = 300,
  lag_setting = 1:4,
  length_val = 16,
  length_test = 8,
  save_model = FALSE,
  filepath = NULL,
  ...
)
}
\arguments{
\item{DT}{univariate time series - data.table object with 2 variables:
"index" (Posixct date-time) and "value". Usually single DT from \code{rsample}}

\item{epochs}{number of epochs to train model (default to 300)}

\item{lag_setting}{numeric vector specifying lags to train on}

\item{length_val}{length of validation set}

\item{length_test}{length of test set}

\item{save_model}{save model? If TRUE, \code{file_path} required}

\item{filepath}{path to save model. Must end with \code{.hdf5}}

\item{...}{additional parameters passed to Keras}
}
\value{
Predictions
}
\description{
Title
}
\examples{
\dontrun{

# Prepare data
apple <- fcf::dow30[
ticker == "AAPL" & date > as.POSIXct("1995-01-01"),
  .SD, .SDcols = c("date", "fcf")
]
data.table::setnames(apple, c("date", "fcf"), c("index", "value"))

# Predict
predict_keras_sequential(DT = apple)

}
}
