% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{prep_financials}
\alias{prep_financials}
\title{Clean up returned list of financial statement data}
\usage{
prep_financials(list_in)
}
\arguments{
\item{list_in}{List returned from \code{\link[=scrape_financials]{scrape_financials()}}.}
}
\description{
Clean up returned list of financial statement data
}
