% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{scrape_financials}
\alias{scrape_financials}
\title{Scrape Financial Statements}
\usage{
scrape_financials(sym)
}
\arguments{
\item{sym}{The stock ticker symbol of interest, passed as a \code{string}.}
}
\description{
Scrape Financial Statements
}
