% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{scrape_shares_outstanding}
\alias{scrape_shares_outstanding}
\title{Get latest outstanding share counts}
\usage{
scrape_shares_outstanding(sym)
}
\arguments{
\item{sym}{The stock ticker symbol of interest, passed as a \code{string}.}
}
\description{
Get latest outstanding share counts
}
