% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_apple}
\alias{ts_apple}
\title{Quarterly reported free cash flow from Apple as ts object from 1995 to 2020.}
\format{
A quarterly ts object from 1995 to 2020
}
\usage{
data(ts_apple)
}
\description{
Quarterly reported free cash flow from Apple as ts object from 1995 to 2020.
}
\keyword{datasets}
