% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_unh}
\alias{ts_unh}
\title{Quarterly reported EBIT from UNH as ts object from 1995 to 2020.}
\format{
A quarterly ts object from 1995 to 2020
}
\usage{
data(ts_unh)
}
\description{
Quarterly reported EBIT from UNH as ts object from 1995 to 2020.
}
\keyword{datasets}
