% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_keras_rnn_eval.R
\name{tune_keras_rnn_eval}
\alias{tune_keras_rnn_eval}
\title{Evaluate (tuned) recurrent neural networks per split/sample}
\usage{
tune_keras_rnn_eval(
  fc_sample,
  cv_setting,
  bayes_best_par,
  col_id = NULL,
  col_date = "index",
  col_value = "value",
  multiple_h = NULL,
  frequency = 4,
  level = 95
)
}
\arguments{
\item{fc_sample}{forecasts per \code{rsample} split}

\item{cv_setting}{list of "periods_train", "periods_val", "periods_test" and
"skip_span" for \link[rsample]{rolling_origin}}

\item{bayes_best_par}{tuned hyperparameters, from \code{tune_keras_rnn_bayesoptim()}}

\item{col_id}{Optional ID column in \code{data}, default to "ticker"}

\item{col_date}{Date column in \code{data}, default to "index"}

\item{col_value}{Value column in \code{data}, default to "value"}

\item{multiple_h}{NULL if forecast horizon equals cv_setting$n_test, else
named list of forecast horizons for accuracy measures}

\item{frequency}{time series frequency, e.g. 4 for quarters and 12 for months}

\item{level}{level for prediction interval in percentage}
}
\value{
list of time series accuracy measures (MAPE, sMAPE, MASE, sMIS, ACD)
optionally for different forecast horizons (if \code{multiple_h} specified)
}
\description{
Evaluate (tuned) recurrent neural networks per split/sample
}
