% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dow30}
\alias{dow30}
\title{US index Dow Jones 30 quarterly earnings data from 1990 to 2020.}
\format{
A data.table with 3600 rows and 9 variables (xz-compressed):
\describe{
\item{company}{Dow Jones listed companies in (August 2020)}
\item{ticker}{bbreviation of company title}
\item{date}{Publishing date of (quarterly) report in POSIX format.
Month and year of interest}
\item{ebit}{Earnings before interest and taxes}
\item{ebitda}{Earnings before intterest, taxes, depreciation and amortization}
\item{net_income}{Net Income}
\item{revenue}{Revenue}
\item{is_adjusted_operating_income}{adjustment}
\item{is_ni_available_to_common_adj}{adjusted net income}
}
}
\usage{
data(dow30)
}
\description{
US index Dow Jones 30 quarterly earnings data from 1990 to 2020.
}
\keyword{datasets}
