% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fc_baseline}
\alias{fc_baseline}
\title{Baseline Forecast (data.table object) with SNAIVE and Holt's Trend}
\format{
A data.table with 34 rows and 3 variables:
\describe{
\item{index}{Publishing date of (quarterly) report}
\item{value}{Earnings before interest and taxes}
\item{type}{Actual or Forecast group}
}
}
\usage{
data(fc_baseline)
}
\description{
Baseline Forecast (data.table object) with SNAIVE and Holt's Trend
}
\keyword{datasets}
