% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_basic_sequential.R
\name{keras_basic_sequential}
\alias{keras_basic_sequential}
\title{Basic sequential NN by Keras framework}
\usage{
keras_basic_sequential(
  X,
  Y,
  return_model = FALSE,
  n_epochs = 200,
  loss = "mse",
  metrics = NULL,
  optimizer_type = "rmsprop",
  dropout = 0.2,
  recurrent_dropout = 0,
  n_units = 32,
  patience = 10,
  live_plot = FALSE
)
}
\arguments{
\item{X}{list of "train", "val", and "test" with 3D (keras) arrays}

\item{Y}{list of "train", "val", and "test" with 2D (keras) arrays}

\item{return_model}{logical, return model?}

\item{n_epochs}{default 200}

\item{loss}{default "mae"}

\item{metrics}{default "mse"}

\item{optimizer_type}{One of "rmsprop" (default) and "adam"}

\item{dropout}{dropout rate}

\item{recurrent_dropout}{Dropout rate applied to reccurent layer. Default 0}

\item{n_units}{32 (currently fixed)}

\item{patience}{when to stop early (default 10)}
}
\value{
evaluation scores for training, validation and test set and model (optional)
}
\description{
Basic sequential NN by Keras framework
}
