% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_sequential.R
\name{keras_sequential}
\alias{keras_sequential}
\title{Train Recurrent Neural Network with Gated Recurrent Unit (GRU) or
Long-Short Term Memory (LSTM) using Keras framework}
\usage{
keras_sequential(
  X,
  Y,
  model_type,
  tsteps,
  return_model = FALSE,
  n_epochs = 200,
  n_units = 32,
  loss = "mae",
  metrics = c("mse"),
  optimizer_type = "rmsprop",
  dropout = 0,
  recurrent_dropout = 0,
  patience = NULL
)
}
\arguments{
\item{X}{list of "train", "val", and "test" with 3D (keras) arrays}

\item{Y}{list of "train", "val", and "test" with 2D (keras) arrays}

\item{model_type}{One of "basic", "gru" and "lstm"}

\item{tsteps}{number of time steps for keras input shape}

\item{return_model}{logical, return model?}

\item{n_epochs}{default 200}

\item{n_units}{32 (currently fixed)}

\item{loss}{default "mae"}

\item{metrics}{default "mse"}

\item{optimizer_type}{One of "rmsprop" (default) and "adam"}

\item{dropout}{dropout rate}

\item{recurrent_dropout}{Dropout rate applied to reccurent layer. Default 0}

\item{patience}{when to stop early (default 10)}
}
\value{
evaluation scores for training, validation and test set
}
\description{
Train Recurrent Neural Network with Gated Recurrent Unit (GRU) or
Long-Short Term Memory (LSTM) using Keras framework
}
