% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_point.R
\name{mase}
\alias{mase}
\title{Mean Absolute Scaled Error (MASE)}
\usage{
mase(data, actual, forecast, m)
}
\arguments{
\item{data}{time series (only train set)}

\item{actual}{actual values (Equivalent of forecast)}

\item{forecast}{predicted values}

\item{m}{frequency, e.g. 12 for monthly and 4 for quarterly series}
}
\value{
numeric vector of length 1
}
\description{
Mean Absolute Scaled Error (MASE), scaled according to M4 Forecasting
competition
}
\examples{
data <- tsRNN::fc_arima[key == "actual", value]
forecast <- tsRNN::fc_arima[key == "predict", value]
data_train <- data[1:(length(data) - length(forecast))]
data_test <- data[(length(data) - length(forecast) + 1):length(data)]

mase(data = data_train, actual = data_test, forecast = forecast, m = 4)

}
\references{
\itemize{
\item 	Svetunkov, I., Sagaert, Y. R. (2020). greybox: Toolbox for Model
Building and Forecasting. \url{https://cran.r-project.org/package=greybox}
\item Makridakis, S., Spiliotis, E., & Assimakopoulos, V. (2020). The M4
Competition: 100,000 time series and 61 forecasting methods. International
Journal of Forecasting, 36(1), 54–74. \url{https://doi.org/10.1016/j.ijforecast.2019.04.014}
}
}
\seealso{
Other PF accuracy measures: 
\code{\link{mape}()},
\code{\link{smape}()}
}
\concept{PF accuracy measures}
