% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baselines.R
\name{plot_baselines}
\alias{plot_baselines}
\title{Plot Forecasts by baseline methods}
\usage{
plot_baselines(
  data,
  title,
  size = 0.6,
  alpha = 0.8,
  colors,
  legend = "bottom",
  scale = NULL
)
}
\arguments{
\item{data}{data.frame containing "index" (Date), "value" (numeric) and type
(character - "Actual" and forecast type)}

\item{title}{diagram title}

\item{size}{line size}

\item{alpha}{line alpha value}

\item{colors}{named character vector to fit color to type}

\item{legend}{legend position}

\item{scale}{NULL or two Date values to scale x-axis}
}
\value{
ggplot2 object
}
\description{
Plot Forecasts by baseline methods
}
