% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baselines.R
\name{plot_baselines_samples}
\alias{plot_baselines_samples}
\title{Plot cross validated samples of forecasts by baseline methods}
\usage{
plot_baselines_samples(
  splits,
  title,
  colors,
  date_type = "datetime",
  ncol = 2,
  scale
)
}
\arguments{
\item{splits}{list of prediction data.frames}

\item{title}{diagram title}

\item{colors}{named character vector to fit color to type}

\item{date_type}{string vector - one of "datetime", "date" or "character"}

\item{ncol}{Number of columns}

\item{scale}{NULL or two Date values to scale x-axis}
}
\value{
ggplot2 object
}
\description{
Plot cross validated samples of forecasts by baseline methods
}
