% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot timeseries and forecast for single split and company}
\usage{
plot_prediction(
  data,
  title,
  size = 0.6,
  alpha = 0.8,
  legend = "bottom",
  scale = NULL,
  PI = FALSE
)
}
\arguments{
\item{data}{data.frame containing "index" (Date), "value" (numeric) and key
(character - "actual" and "predict")}

\item{title}{plot title}

\item{size}{line size}

\item{alpha}{line alpha value}

\item{legend}{legend position}

\item{scale}{NULL or two Date values to scale x-axis}

\item{PI}{add prediction interval?}
}
\value{
ggplot2 object
}
\description{
Plot timeseries and forecast for single split and company
}
