% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction_samples}
\alias{plot_prediction_samples}
\title{Plot multiple splits from list with forecast results}
\usage{
plot_prediction_samples(
  splits,
  col_date = "index",
  col_value = "value",
  col_group = "key",
  title = NULL,
  date_type = "datetime",
  ncol = 3L,
  scale = NULL,
  PI = FALSE,
  col_pi_high = "hi95",
  col_pi_low = "lo95"
)
}
\arguments{
\item{splits}{list of prediction data.frames}

\item{col_date}{Date column, default to "index"}

\item{col_value}{Value column, default to "value"}

\item{col_group}{Group column , default to "key" ("actual" and
"predict") - see example}

\item{title}{diagram title}

\item{date_type}{string vector - one of "datetime", "date" or "character"}

\item{ncol}{Number of columns}

\item{scale}{NULL or two Date values to scale x-axis}

\item{PI}{add prediction interval?}

\item{col_pi_high}{Upper confidence bound}

\item{col_pi_low}{Lower confidence bound}
}
\value{
ggplot2 object
}
\description{
Plot multiple splits from list with forecast results
}
\examples{
data <- tsRNN::fc_arima
plot_prediction_samples(
  splits = list(data, data),
  ncol = 2L
)

# with Prediction Interval
plot_prediction_samples(
  splits = list(data, data),
  ncol = 2L,
  PI = TRUE
)
}
