% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction_samples}
\alias{plot_prediction_samples}
\title{Plot multiple splits from same \code{rsample} split}
\usage{
plot_prediction_samples(splits, title, date_type = "datetime", ncol = 2, scale)
}
\arguments{
\item{splits}{list of prediction data.frames}

\item{title}{diagram title}

\item{date_type}{string vector - one of "datetime", "date" or "character"}

\item{ncol}{Number of columns}

\item{scale}{NULL or two Date values to scale x-axis}
}
\value{
ggplot2 object
}
\description{
Plot multiple splits from same \code{rsample} split
}
