% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_arima.R
\name{predict_arima}
\alias{predict_arima}
\title{Forecast the next n steps by ARIMA for each split}
\usage{
predict_arima(data, cv_setting, transform = NULL)
}
\arguments{
\item{data}{data.table object}

\item{cv_setting}{cross valudation settings. List Requiring \itemize{
\item periods_train
\item periods_val
\item periods_test
\item skip_span
}}

\item{transform}{One of NULL (default) and "normalize"}
}
\value{
list with accuracy and forecasts (point forecast and PI)
}
\description{
Forecast the next n steps by ARIMA for each split
}
