% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_baselines.R
\name{predict_baselines}
\alias{predict_baselines}
\title{Prediction and evaluation for baseline models including cross validation}
\usage{
predict_baselines(data, cv_setting, normalize)
}
\arguments{
\item{data}{data.table object}

\item{cv_setting}{cross valudation settings. List Requiring \itemize{
\item periods_train
\item periods_val
\item periods_test
\item skip_span
}}

\item{normalize}{normalize data?}
}
\value{
list with accuracy and forecasts (point forecast and PI)
}
\description{
Prediction and evaluation for baseline models including cross validation
}
