% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_baselines.R
\name{predict_baselines}
\alias{predict_baselines}
\title{Prediction and evaluation for baseline models including cross validation}
\usage{
predict_baselines(
  data,
  cv_setting,
  col_id = NULL,
  col_date = "index",
  col_value = "value",
  transform = NULL,
  frequency = 4,
  multiple_h = NULL
)
}
\arguments{
\item{data}{data.table object}

\item{cv_setting}{cross valudation settings. List Requiring \itemize{
\item periods_train
\item periods_val
\item periods_test
\item skip_span
}}

\item{col_id}{Optional ID column in \code{data}, default to "ticker"}

\item{col_date}{Date column in \code{data}, default to "index"}

\item{col_value}{Value column in \code{data}, default to "value"}

\item{transform}{One of NULL, "normalize" and "box" for Box-Cox transformation}

\item{frequency}{time series frequency, e.g. 4 for quarters and 12 for months}

\item{multiple_h}{NULL if forecast horizon equals cv_setting$n_test, else
named list of forecast horizons for accuracy measures}
}
\value{
list with accuracy and forecasts (point forecast and PI)
}
\description{
Prediction and evaluation for baseline models including cross validation
}
