% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_dropout_model.R
\name{py_dropout_model}
\alias{py_dropout_model}
\title{Change dropout rate in recurrent layer or dropout layer of trained model}
\usage{
py_dropout_model(model, dropout = 0.1)
}
\arguments{
\item{model}{Keras model, required}

\item{dropout}{new dropout rate}
}
\value{
Keras model
}
\description{
Wrapper for python function. Model requires fixed training mode (else dropout
turned off automatically during testing and changing dropout rate has no
effect)
}
\details{
Currently requires python warnings wrapper to suppress tensorflow
warning
}
\examples{
# Load Model with permanently activated dropout and generate test data
model <- keras::load_model_hdf5(
  system.file("tinytest_data/simple_rnn_dropout.hdf5", package = "tsRNN")
)
test <- array(runif(18, 0.1, 3), dim = c(6L, 2L, 1L))

# 1. Deactivate dropout for testing
model_01 <- py_dropout_model(model, 0)
model_01(test)

# 2. Change dropout level
model_02 <- py_dropout_model(model, 0.8)
model_02(test)
}
