% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_nn_preparation.R
\name{ts_nn_preparation}
\alias{ts_nn_preparation}
\title{Timeseries data preparation for neural network Keras models}
\usage{
ts_nn_preparation(data, lag_setting = 1:4, length_val = 16, length_test = 8)
}
\arguments{
\item{data}{balanced timeseries data}

\item{lag_setting}{by how much is y lagged}

\item{length_val}{length of validation set}

\item{length_test}{length of test set}
}
\value{
list of "X" and "Y" each containing "train", "val" and "test" arrays
}
\description{
The following steps are proceeded:
\itemize{
\item New columns "value_lagX", with X for #lags, specified by \code{lag_setting}
\item Split in train, validation and test split
\item Transformation to 3D arrays for Keras
}
}
\examples{

}
