% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_normalization.R
\name{ts_normalization}
\alias{ts_normalization}
\title{Normalize univariate timeseries}
\usage{
ts_normalization(DT, length_val, length_test, metrics = FALSE)
}
\arguments{
\item{DT}{data.table object with columns "index" and "value"}

\item{length_val}{length for validation set}

\item{length_test}{length for test set}

\item{metrics}{return data only or list of data and metrics?}
}
\value{
Depending on \code{metrics}, processed DT object or list of "data" and
"metrics" (center and scale)
}
\description{
Normalize univariate timeseries
}
