% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_keras_rnn_predict.R
\name{tune_keras_rnn_predict}
\alias{tune_keras_rnn_predict}
\title{Use tuned recurrent neural network parameters with Keras functional API and
to train best performing model(s) and forecast}
\usage{
tune_keras_rnn_predict(
  data,
  model_type,
  cv_setting,
  bayes_best_par,
  col_id = NULL,
  col_date = "index",
  col_value = "value",
  level = 95,
  iter = 10,
  iter_dropout = 1000,
  save_model = NULL,
  save_model_id = NULL
)
}
\arguments{
\item{data}{Univariate time series (data.frame) with date and value column,
specified in \code{col_date} and \code{col_value}}

\item{model_type}{One of "simple", "gru" or "lstm"}

\item{cv_setting}{list of "periods_train", "periods_val", "periods_test" and
"skip_span" for \link[rsample]{rolling_origin}}

\item{bayes_best_par}{tuned hyperparameters, from \code{tune_keras_rnn_bayesoptim()}}

\item{col_id}{Optional ID column in \code{data}, default to "ticker"}

\item{col_date}{Date column in \code{data}, default to "index"}

\item{col_value}{Value column in \code{data}, default to "value"}

\item{level}{level for prediction interval in percentage}

\item{iter}{number of neural networks to train per split with same
hyperparameters}

\item{iter_dropout}{number of iterations for prediction intervals calculated
by monte carlo dropout}

\item{save_model}{Automatically save tuned models? Specify NULL for No or
character vector with path to directory for yes}

\item{save_model_id}{optional id for model filename}
}
\value{
list of forecasts per split
}
\description{
Use tuned recurrent neural network parameters with Keras functional API and
to train best performing model(s) and forecast
}
