% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_keras_sequential.R
\name{tune_keras_sequential}
\alias{tune_keras_sequential}
\title{Tune keras basic model with Bayes Optimization and select best performing
model}
\usage{
tune_keras_sequential(data, model_type, cv_setting, tuning_bounds)
}
\arguments{
\item{data}{Univariate time series (data.frame) with columns index and value}

\item{model_type}{One of "basic", "gru" or "lstm"}

\item{cv_setting}{list of "periods_train", "periods_val", "periods_test" and
"skip_span" for \code{rsample}}

\item{tuning_grid}{list of "lags" "optimizer", "dropout"}
}
\value{
list of "results" and "min_params"
}
\description{
Tune keras basic model with Bayes Optimization and select best performing
model
}
\section{Tuning Grid}{

The following parameters are (currently) available for tuning.
\itemize{
\item \code{lags}: named list of numeric vectors with length of lags to use
\item \code{optimizer}: character vector specifying optimizer to use. One of "adam"
and "rmsprop"
\item \code{dropout}: numeric vector specifying dropout rates
}
}

